Change minimap scale types to double

+++ src/minimap_initializer.c
@@ -6,7 +6,7 @@
 /*   By: kwang <kwang@student.42kl.edu.my>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/12 16:16:20 by kwang             #+#    #+#             */
-/*   Updated: 2022/12/12 23:15:46 by kwang            ###   ########.fr       */
+/*   Updated: 2022/12/13 12:55:13 by kwang            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -64,11 +64,11 @@ Returns nothing
 */
 void	cache_minimap(t_map map, void *mlx, t_cache *cache)
 {
-	const int	minimap_size = WIN_HEIGHT / MINIMAP_SCALE;
-	const int	minimap_grid_size = minimap_size / MINIMAP_GRID_SCALE;
-	int			x;
-	int			y;
-	t_data		*asset;
+	const double	minimap_size = WIN_HEIGHT / MINIMAP_SCALE;
+	const double	minimap_grid_size = minimap_size / MINIMAP_GRID_SCALE;
+	int				x;
+	int				y;
+	t_data			*asset;
 
 	create_image(mlx, &cache->minimap,
 		map.width * minimap_grid_size, map.height * minimap_grid_size);
